src/utils/cuopt_gemini_optimizer.py
"""
Real NVIDIA cuOpt + Gemini Integration for Company Optimization

This module uses your actual NVIDIA cuOpt Cloud API key for portfolio optimization
with Google Gemini explanations.
"""

import numpy as np
import pandas as pd
import json
import asyncio
from typing import Dict, List, Any, Optional, Tuple
from datetime import datetime
import logging
import os

# Load environment variables
try:
    from dotenv import load_dotenv
    load_dotenv()
except ImportError:
    pass

# NVIDIA cuOpt Cloud API integration
from .nvidia_cuopt_cloud_api import NVIDIACuOptCloudAPI

# Google Gemini integration
try:
    import google.generativeai as genai
    GEMINI_AVAILABLE = True
except ImportError:
    GEMINI_AVAILABLE = False

class CuOptGeminiOptimizer:
    """
    Real NVIDIA cuOpt + Gemini Optimizer using your official API key
    """
    
    def __init__(self, config_path: str = None):
        """Initialize the optimizer with real NVIDIA cuOpt Cloud API"""
        self.logger = logging.getLogger(__name__)
        
        # Initialize NVIDIA cuOpt Cloud API
        self.nvidia_cuopt = NVIDIACuOptCloudAPI()
        
        # Initialize Gemini for additional explanations
        self.gemini_api_key = os.getenv('GEMINI_API_KEY')
        if GEMINI_AVAILABLE and self.gemini_api_key:
            genai.configure(api_key=self.gemini_api_key)
            self.gemini_model = genai.GenerativeModel('gemini-1.5-flash')
            self.gemini_available = True
            self.logger.info("✅ Google Gemini initialized for enhanced explanations")
        else:
            self.gemini_available = False
            self.logger.warning("⚠️ Gemini API not available")
        
        self.logger.info("✅ CuOpt + Gemini Optimizer initialized with real NVIDIA cuOpt Cloud API")
    
    async def optimize_and_rank_companies(self, companies: List[Dict[str, Any]], 
                                        criteria: Dict[str, Any],
                                        top_n: int = 10) -> Dict[str, Any]:
        """
        Main function using real NVIDIA cuOpt for portfolio optimization
        
        Args:
            companies: List of companies from search engine
            criteria: User search criteria
            top_n: Number of top companies to return (1-10)
            
        Returns:
            Dictionary with NVIDIA cuOpt optimized companies and AI explanations
        """
        try:
            self.logger.info(f"🚀 Starting real NVIDIA cuOpt optimization for {len(companies)} companies, selecting top {top_n}")
            
            # Use your NVIDIA cuOpt Cloud API
            cuopt_result = await self.nvidia_cuopt.optimize_portfolio_with_cuopt(
                companies, criteria, top_n
            )
            
            if cuopt_result['success']:
                self.logger.info(f"✅ NVIDIA cuOpt optimization completed successfully")
                self.logger.info(f"📊 Method: {cuopt_result['optimization_method']}")
                self.logger.info(f"🏆 Selected {cuopt_result['selected_count']} companies")
                self.logger.info(f"⏱️  Optimization time: {cuopt_result.get('optimization_time', 0):.2f}s")
                
                return cuopt_result
            else:
                self.logger.warning("NVIDIA cuOpt optimization unsuccessful, using mathematical fallback")
                return await self._create_fallback_results(companies, criteria, top_n)
            
        except Exception as e:
            self.logger.error(f"❌ Error in NVIDIA cuOpt optimization process: {e}")
            return await self._create_fallback_results(companies, criteria, top_n)
    
    async def _create_fallback_results(self, companies: List[Dict[str, Any]], 
                                     criteria: Dict[str, Any], 
                                     top_n: int) -> Dict[str, Any]:
        """Create mathematical optimization fallback results"""
        self.logger.info("📊 Creating advanced mathematical optimization fallback")
        
        # Advanced mathematical optimization fallback
        try:
            import scipy.optimize as opt
            
            # Prepare optimization data
            n_companies = len(companies)
            spr_scores = np.array([float(c.get('enhanced_spr_score', 0)) for c in companies])
            
            # Risk scores
            risk_scores = []
            for company in companies:
                pe_ratio = float(company.get('pe_ratio', 15))
                debt_ratio = float(company.get('debt_to_equity', 0.5))
                market_cap = float(company.get('market_cap', 1e9))
                
                # Calculate risk
                pe_risk = max(0, (pe_ratio - 15) / 30)
                debt_risk = min(1, debt_ratio / 2)
                size_risk = max(0, (1e12 - market_cap) / 1e12)
                risk_score = (pe_risk + debt_risk + size_risk) / 3
                risk_scores.append(risk_score)
            
            risk_scores = np.array(risk_scores)
            
            # Objective function: maximize SPR while managing risk
            def objective(x):
                selected_spr = np.sum(x * spr_scores)
                selected_risk = np.sum(x * risk_scores)
                
                # Risk penalty based on criteria
                risk_tolerance = criteria.get('risk_tolerance', 'Moderate')
                if risk_tolerance == 'Conservative':
                    risk_penalty = selected_risk * 3
                elif risk_tolerance == 'Moderate':
                    risk_penalty = selected_risk * 1.5
                else:  # Aggressive
                    risk_penalty = selected_risk * 0.8
                
                return -(selected_spr - risk_penalty)  # Negative for minimization
            
            # Constraints
            constraints = [
                {'type': 'eq', 'fun': lambda x: np.sum(x) - top_n},  # Select exactly top_n
            ]
            
            # Bounds (binary variables approximated as [0,1])
            bounds = [(0, 1) for _ in range(n_companies)]
            
            # Initial guess
            x0 = np.zeros(n_companies)
            top_indices = np.argsort(spr_scores)[-top_n:]
            x0[top_indices] = 1
            
            # Solve optimization
            result = opt.minimize(
                objective, x0, 
                method='SLSQP',
                bounds=bounds,
                constraints=constraints,
                options={'maxiter': 1000, 'ftol': 1e-6}
            )
            
            # Convert to binary solution
            solution = np.round(result.x)
            selected_indices = np.where(solution > 0.5)[0]
            
            # Build selected companies
            selected_companies = []
            for rank, idx in enumerate(selected_indices, 1):
                company = companies[idx].copy()
                company['cuopt_rank'] = rank
                company['final_spr_score'] = company.get('enhanced_spr_score', 0)
                company['optimization_score'] = spr_scores[idx]
                company['risk_score'] = risk_scores[idx]
                company['selection_reason'] = f"Selected by mathematical optimization (rank: {rank})"
                selected_companies.append(company)
            
            # Sort by optimization score
            selected_companies.sort(key=lambda x: x['optimization_score'], reverse=True)
            
            # Update ranks after sorting
            for rank, company in enumerate(selected_companies, 1):
                company['cuopt_rank'] = rank
            
            optimization_method = "Advanced Mathematical Optimization (cuOpt Fallback)"
            objective_value = -result.fun if result.success else 0
            
        except Exception as e:
            self.logger.error(f"Mathematical optimization failed: {e}")
            # Simple SPR-based fallback
            scored_companies = []
            for company in companies:
                spr_score = float(company.get('enhanced_spr_score', 0))
                company_copy = company.copy()
                company_copy['final_spr_score'] = spr_score
                company_copy['optimization_score'] = spr_score
                company_copy['selection_reason'] = f"Selected by SPR ranking"
                scored_companies.append(company_copy)
            
            # Sort and select top N
            scored_companies.sort(key=lambda x: x['optimization_score'], reverse=True)
            selected_companies = scored_companies[:top_n]
            
            # Add rankings
            for rank, company in enumerate(selected_companies, 1):
                company['cuopt_rank'] = rank
            
            optimization_method = "SPR Ranking (Simple Fallback)"
            objective_value = sum(c['optimization_score'] for c in selected_companies)
        
        # Generate explanations
        if self.gemini_available:
            try:
                explanations = await self._generate_fallback_explanations(selected_companies, criteria, optimization_method)
            except Exception as e:
                self.logger.error(f"Explanation generation failed: {e}")
                explanations = f"Portfolio optimized using {optimization_method}. Mathematical selection completed successfully."
        else:
            explanations = f"Portfolio optimized using {optimization_method}. Mathematical selection completed successfully."
        
        return {
            'success': True,
            'top_companies': selected_companies,
            'ai_explanations': {'overall_explanation': explanations},
            'optimization_method': optimization_method,
            'total_candidates': len(companies),
            'selected_count': len(selected_companies),
            'solver_status': 'completed',
            'optimization_time': 0.5,
            'objective_value': objective_value
        }
    
    async def _generate_fallback_explanations(self, companies: List[Dict], criteria: Dict, method: str) -> str:
        """Generate explanations for fallback optimization"""
        try:
            company_info = []
            for company in companies[:5]:
                info = f"{company.get('symbol', 'N/A')} (SPR: {company.get('final_spr_score', 0):.2f}, Risk: {company.get('risk_score', 0):.2f})"
                company_info.append(info)
            
            prompt = f"""
            Explain this mathematically optimized investment portfolio:
            
            Top 5 Selected Companies: {', '.join(company_info)}
            Optimization Method: {method}
            Selection Criteria: {criteria}
            Total Companies: {len(companies)}
            
            Provide a professional investment analysis covering:
            1. Mathematical optimization methodology used
            2. SPR-based selection rationale and benefits
            3. Portfolio balance, risk management, and diversification
            4. Investment recommendations and outlook
            5. Why this mathematical approach is superior to manual selection
            
            Focus on the scientific approach and mathematical precision of the selection process.
            """
            
            response = self.gemini_model.generate_content(prompt)
            return response.text
            
        except Exception as e:
            self.logger.error(f"Fallback explanation generation failed: {e}")
            return f"Portfolio optimized using {method}. Advanced mathematical algorithms were used to balance SPR scores with risk considerations for optimal investment selection."
