requirements.txt
# Optimized requirements for Vercel deployment
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
plotly>=5.14.0
dash>=2.10.0
dash-bootstrap-components>=1.4.0

# AI and ML (lightweight)
google-generativeai>=0.3.0
mistralai>=0.4.2
scikit-learn>=1.3.0
textblob>=0.17.0

# Data processing and APIs
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
aiohttp>=3.9.1

# Financial data APIs
yfinance>=0.2.0
alpha-vantage>=2.3.0
finnhub-python>=2.4.19
newsapi-python>=0.2.7

# Configuration and utilities
python-dotenv>=1.0.0
pyyaml>=6.0

# API and web framework
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.2.0

# Data export/import
openpyxl>=3.1.0

# Date and time handling
python-dateutil>=2.8.0

# Additional dependencies
httpx>=0.25.0